requirements.txt
Flask==1.1.2
requests==2.23.0
PyJWT==1.7.1
pymssql==2.1.4
